+++ CHANGELOG.md
@@ -13,10 +13,12 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
-  
+
   ### :rocket: New Feature
   - `preview`
     - [#435](https://github.com/wix-incubator/rich-content/pull/435) package added
+  - `emoji`
+    - plugin rewrite
   ### :bug: Bug Fix
   - `common`
     - [#458](https://github.com/wix-incubator/rich-content/pull/458) normalization: sparse entity map keys handled correctly